requirements.txt
NearBeach
django-probes
django-environ